Time: 12 ms (6.92%) | Memory: 106.2 MB (5.61%) - LeetSync

+++ 2358-number-of-ways-to-split-array/number-of-ways-to-split-array.cpp
@@ -3,15 +3,15 @@ public:
     int waysToSplitArray(vector<int>& nums) {
         int n  = nums.size();
 
-        vector<long> prefix = {nums[0]};
+        vector<long long> prefix = {nums[0]};
         for(int i = 1; i<n; ++i){
             prefix.push_back(nums[i] + prefix.back());
         }
 
         int ans = 0;
         for(int i = 0; i<n-1; ++i){
-            long leftSide = prefix[i];
-            long rightSide = prefix.back()-prefix[i];
+            long long leftSide = prefix[i];
+            long long rightSide = prefix.back()-prefix[i];
             if(leftSide>=rightSide){
                 ++ans;
             }
